fix(pages/home): only load font picker on client
